Remove secp256k1-node
