+++ Cargo.toml
@@ -1,9 +1,22 @@
-[workspace]
-members = [
-    "secp256k1",
-    "secp256k1-node",
-    "secp256k1-wasm",
-]
+[package]
+name = "secp256k1-wasm"
+version = "0.0.0"
+authors = ["Kirill Fomichev <fanatid@ya.ru>"]
+edition = "2018"
+description = "A Rust library for building tiny-secp256k1 WASM."
+license = "MIT"
+publish = false
+
+[lib]
+crate-type = ["cdylib"]
+
+[dependencies]
+# `[patch.crates-io]` is not working :(
+# `more-features` branch add features: ecdh, extrakeys, schnorrsig.
+# This reuqired because LTO do not strip not used functions in resulted WASM file.
+# secp256k1-sys PR: https://github.com/rust-bitcoin/rust-secp256k1/pull/287
+secp256k1-sys = { version = "0.4.0", default-features = false, git = "https://github.com/TheBlueMatt/rust-secp256k1", branch = "2021-03-no-vis" }
+# secp256k1-sys = { version = "0.4.0", default-features = false, git = "https://github.com/fanatid/rust-secp256k1", branch = "more-features" }
 
 [profile.release]
 lto = true
