+++ Makefile
@@ -2,38 +2,15 @@
 build-js:
 	npx tsc
 
-.PHONY: build-addon-%
-build-addon-%: export PAIR = $(subst +, ,$(subst build-addon-,,$@))
-build-addon-%:
-	$(if $(findstring musl,$(firstword $(PAIR))),RUSTFLAGS="-C target-feature=-crt-static",) cargo build \
-		--package secp256k1-node \
-		--target $(firstword $(PAIR)) \
-		$(if $(findstring musl,$(firstword $(PAIR))),,-Z build-std=panic_abort,std) \
-		--release
-	mkdir -p lib && cp -f target/$(firstword $(PAIR))/release/`node util/addon-target-name.js` lib/secp256k1-$(lastword $(PAIR))
-	strip $(if $(findstring Darwin,$(shell uname -s)),-Sx,--strip-all) lib/secp256k1-$(lastword $(PAIR))
-
-.PHONY: build-addon-debug
-build-addon-debug:
-	cargo build --package secp256k1-node
-
-.PHONY: build-addon-debug-%
-build-addon-debug-%: export PAIR = $(subst +, ,$(subst build-addon-debug-,,$@))
-build-addon-debug-%:
-	cargo build --package secp256k1-node --target $(firstword $(PAIR))
-	mkdir -p lib && cp -f target/$(firstword $(PAIR))/debug/`node util/addon-target-name.js` lib/secp256k1-$(lastword $(PAIR))
-
 .PHONY: build-wasm
 build-wasm:
-	RUSTFLAGS="-C link-args=-zstack-size=655360" cargo build --package secp256k1-wasm --target wasm32-unknown-unknown --release
+	RUSTFLAGS="-C link-args=-zstack-size=655360" cargo build --target wasm32-unknown-unknown --release
 	mkdir -p lib && cp -f target/wasm32-unknown-unknown/release/secp256k1_wasm.wasm lib/secp256k1.wasm
-	wasm-opt --strip-debug --strip-producers --output lib/secp256k1.wasm lib/secp256k1.wasm
-	node util/wasm-strip.js lib/secp256k1.wasm
-	wasm-opt -O4 --output lib/secp256k1.wasm lib/secp256k1.wasm
+	wasm-opt -O4 --strip-debug --strip-producers --output lib/secp256k1.wasm lib/secp256k1.wasm
 
 .PHONY: build-wasm-debug
 build-wasm-debug:
-	RUSTFLAGS="-C link-args=-zstack-size=655360" cargo build --package secp256k1-wasm --target wasm32-unknown-unknown
+	RUSTFLAGS="-C link-args=-zstack-size=655360" cargo build --target wasm32-unknown-unknown
 	mkdir -p lib && cp -f target/wasm32-unknown-unknown/debug/secp256k1_wasm.wasm lib/secp256k1.wasm
 
 .PHONY: clean
@@ -53,7 +30,7 @@ clean:
 		tests/browser \
 		types
 
-eslint_files = benches/*.{js,json} examples/**/*.{js,json} src_ts/*.ts tests/*.js util/*.js *.json *.cjs
+eslint_files = benches/*.{js,json} examples/**/*.{js,json} src_ts/*.ts tests/*.js *.json *.cjs
 
 .PHONY: format
 format:
@@ -62,14 +39,12 @@ format:
 	npx sort-package-json \
 		package.json \
 		benches/package.json \
-		examples/random-in-node/package.json \
-		examples/react-app/package.json
+		examples/*/package.json
 
 .PHONY: lint
 lint:
 	cargo fmt -- --check
-	cargo clippy --package secp256k1-node
-	cargo clippy --package secp256k1-wasm --target wasm32-unknown-unknown
+	cargo clippy --target wasm32-unknown-unknown
 	npx eslint $(eslint_files)
 
 .PHONY: test
@@ -106,7 +81,7 @@ test-node-raw-ci:
 	$(test_node_raw) --no-ansi --no-progress
 
 .PHONY: test-node
-test-node: build-js build-addon-debug build-wasm-debug test-node-raw
+test-node: build-js build-wasm-debug test-node-raw
 
 .PHONY: test-node-coverage-raw
 test-node-coverage-raw:
