+++ package.json
@@ -1,7 +1,6 @@
 {
   "name": "tiny-secp256k1",
   "version": "1.1.6",
-  "private": true,
   "description": "A tiny secp256k1 JS",
   "homepage": "https://github.com/bitcoinjs/tiny-secp256k1#readme",
   "bugs": {
@@ -15,16 +14,17 @@
   "type": "module",
   "main": "./lib/index.js",
   "browser": {
-    "./lib/index.js": "./lib/index.browser.js",
     "./lib/rand.js": "./lib/rand.browser.js",
     "./lib/wasm_loader.js": "./lib/wasm_loader.browser.js"
   },
   "types": "./types/index.d.ts",
+  "files": [
+    "lib"
+  ],
   "devDependencies": {
     "@types/node": "^14.14.35",
     "@typescript-eslint/eslint-plugin": "^4.18.0",
     "@typescript-eslint/parser": "^4.18.0",
-    "binaryen": "^100.0.0",
     "browser-run": "^10.0.0",
     "buffer": "^6.0.3",
     "eslint": "^7.22.0",
