--- secp256k1-node/Cargo.toml
@@ -1,20 +1,0 @@
-[package]
-name = "secp256k1-node"
-version = "0.0.0"
-authors = ["Kirill Fomichev <fanatid@ya.ru>"]
-edition = "2018"
-description = "A Rust library for building tiny-secp256k1 Node.js addon."
-license = "MIT"
-publish = false
-
-[lib]
-crate-type = ["cdylib"]
-
-[dependencies]
-lazy_static = "1.4.0"
-napi = { version = "1.3.2", default-features = false, features = ["napi3"] }
-napi-derive = "1.0"
-secp256k1 = { path = "../secp256k1", features = ["std"] }
-
-[build-dependencies]
-napi-build = "1.0"
