--- secp256k1-wasm/Cargo.toml
@@ -1,14 +1,0 @@
-[package]
-name = "secp256k1-wasm"
-version = "0.0.0"
-authors = ["Kirill Fomichev <fanatid@ya.ru>"]
-edition = "2018"
-description = "A Rust library for building tiny-secp256k1 WASM."
-license = "MIT"
-publish = false
-
-[lib]
-crate-type = ["cdylib"]
-
-[dependencies]
-secp256k1 = { path = "../secp256k1" }
