--- secp256k1/Cargo.toml
@@ -1,18 +1,0 @@
-[package]
-name = "secp256k1"
-version = "0.0.0"
-authors = ["Kirill Fomichev <fanatid@ya.ru>"]
-edition = "2018"
-description = "A secp256k1-sys wrapper for building tiny-secp256k1."
-license = "MIT"
-publish = false
-
-[dependencies]
-# `[patch.crates-io]` is not working :(
-# `more-features` branch add features: ecdh, extrakeys, schnorrsig.
-# This reuqired because LTO do not strip not used functions in resulted WASM file.
-# secp256k1-sys PR: https://github.com/rust-bitcoin/rust-secp256k1/pull/287
-secp256k1-sys = { version = "0.4.0", default-features = false, git = "https://github.com/fanatid/rust-secp256k1", branch = "more-features" }
-
-[features]
-std = []
