+++ tests/index.js
@@ -12,39 +12,6 @@ test.onFinish(() => {
   }
 });
 
-function runTests(secp256k1) {
-  if (secp256k1 !== null) {
-    test_ecdsa(secp256k1);
-    test_points(secp256k1);
-    test_privates(secp256k1);
-  }
-}
-
-runTests(secp256k1.__addon);
-runTests(secp256k1.__wasm);
-
-test("functions exported properly", (t) => {
-  const fnList = [
-    "isPoint",
-    "isPointCompressed",
-    "isPrivate",
-    "pointAdd",
-    "pointAddScalar",
-    "pointCompress",
-    "pointFromScalar",
-    "pointMultiply",
-    "privateAdd",
-    "privateSub",
-    "sign",
-    "verify",
-  ];
-  const source =
-    secp256k1.__initializeContext === secp256k1.__wasm.__initializeContext
-      ? secp256k1.__wasm
-      : secp256k1.__addon;
-  for (const fnName of fnList) {
-    t.same(secp256k1[fnName], source[fnName]);
-  }
-
-  t.end();
-});
+test_ecdsa(secp256k1);
+test_points(secp256k1);
+test_privates(secp256k1);
